backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
crewai==0.11.0
pydantic==2.5.2
python-multipart==0.0.6
httpx<0.25.0,>=0.24.0
openai==1.7.1
langchain>=0.1.0,<0.2.0
langchain-openai==0.0.2.post1 